Rely more on superusers for upper-level tasks
The ability for arbitrary to modify user permissions is inherently
dangerous and should only be allowed by superusers until a better
solution is found.

The ability to imitate other users is also quite dangerous and should be
left only to superusers.

The ability to add and change categories is significantly rare and
should be left only to superusers.

We can add a management command that creates superusers though. So, you
know, that's nice.