Improve performance and add fields to user report
Django's `.annotate()` function that we were using in the User Report
causes a "GROUP BY" on **EVERY** column being selected. This freezes up
the page load.

If we instead do subqueries using `.extra()` we can let the database
optimize for us. Django's documentation says that `.extra()` will be
depreciated, but it DOESN'T say that `.annotate()` will GROUP BY every
column, so until it says "this is depreciated" we're going with it.