Stop using AuthGroup for group owner permissions
Previously we would create a new Django `AuthGroup` which had
permissions attached to it such as the ability to send direct messages
to all users. This was hard-coded with an OFA-specific name so to make
things less OFA-specific (and further reduce the reliance on `AuthGroup`
we can just add the permissions individually.

Note: This won't really handle things if a user is removed from being a
group owner. They'll still have the "send direct messages to everyone"
permission until that's actually removed. Probably a TODO, but not a
huge one since that permission is mostly just to prevent spammers from
abusing the Direct Message functionality. And you don't usually promote
spammers to group owners.