Handle changes that were causing new migrations
There was a typo in the group model `featured` help text, and we changed
the storage for images, which were causing Django to think that new
migrations were needed. Neither is an actual schema change and there are
no 3rd party users of Connect yet, so changing it in old migrations
shouldn't be a huge issue.