Adding `CONNECT_APP` setting, for custom Connect versions
A key feature of Connect is the ability to have multiple custom branding
and functions. Setting the `CONNECT_APP` setting adds that application
to `INSTALLED_APPS` and also tells django's static files finders to
prioritize the static files located in that app.

We hard-code the static file directory in `storage_settings.py` because
Django's "AppDirectoriesFinder" staticfiles finder goes through applications
in the order they are listed in INSTALLED_APPS, and then prioritizes the
FIRST static file it finds which matches a static file lookup.
Django's ``django.template.loaders.app_directories.Loader`` template loader
will go through all the template folders in INSTALLED_APPS and prioritize
the FINAL template it finds which matches a lookup.
Thus we need to hard code in the actual static files we want to override
the open source version with.