Update login_required middleware
Currently AJAX requests were being sent through the login flow, which
could cause problems if you had multiple windows open and active AJAX
requests going.

For example, if you had one window open which was checking for new
messages and you logged out in another window, the next check for new
messages would attempt to go through the login flow. As the login flow
keeps track of the URL you're attempting to access when you first start
the flow it's possible that you could end up being redirected to the
check-for-new-messages JSON endpoint instead of an actual HTTP page.

This also adds no-cache headers to non-logged-in sessions. The headers
added should cover almost all cases of client-side caching on all
browsers.

This also cleans up the tests for the middleware quite a bit. As this
code predates some of the more modern testing mixins that are included
with Connect, this was an opportunity to use some of those to make the
tests slightly less likely to cause problems later as we change Connect.