Fix daily digest notifications
Previously Connect was grossly inefficient in how it handled daily
notification task generation. Not only would it execute the same "send
emails" task multiple times, it would do individual `UPDATE` queries
for every person with a digest. This caused the database to lock-up

This update ensures that the task is only created once per user with a
digest, and significant reduces the number of queries involved.

Still something well worth keeping an eye on.