Upgrading redactor, adding newlines to improve copy/paste
$(elem).text() in jQuery will, by default, remove all of the tags in a
piece of content. What it will not do, is turn <br> into newlines. This
causes a problem with redactor, which when trying to process incoming
text will use similar methods to generate a plain version of the content
being copied in. Causing what people see as multiple lines in their
browser to become a huge mess in the WYSIWYG.

As our "Clean Up Submitted HTML" functionality in the backend removes
all newlines, causing entire messages to appear in 1 line of code, this
causes problems with the "paste" functionality in redactor.

This change upgrades redactor, has content pasted into redactor turned
into plaintext with no formatting when first pasted, and changes the
HTML cleanup functionality in the backend to add newlines after HTML
breaks.