Allow admins to change staff status of users
Up to this point if you wanted to change the `Staff Member` status of an
individual user you needed to go into the database or use the python
shell. This code adds that ability to the "Update User" panel for admins
who are explicitly given permission to update staff status.