Allow relevant users to see autocomplete user list
Previously only staff and superadmins could use the drop-down user
autocomplete service that connect relies on for areas such as group
editing and inviting of new users. This presented a problem because
it's possible for non-staff non-superadmins to have permission to access
those things.

I expanded permission to access the autocomplete to users with
permission to access areas of the app where the autocomplete is used, as
well as staff members and superusers (who will keep access because they
pass all permission tests by default in django)

As we don't want non-staff to have access to email addresses we had to
remove email address from the drop-down. However, it's still possible to
search for email. I'm open to removing that search functionality if we
see that as a major security issue.