Fix redirect when admin edits another user (#34)
Currently if you edit a user's profile it redirects you to your own user
profile. This isn't a huge deal for individual users, but for admins it
is a weird experience.

This adjusts things so that when you update any user you go to that
user's profile page instead of your own.