Hide banned user's replies
We currently do not hide replies that are sent from banned users from
being viewed in the app. Since we want to retain our "shadow ban"
functionality adding a test to the "is this message viewable by {user}"
method that hides messages sent from banned users should address this
issue.