(Re-)Add pagination to group member list page
The Django Pure Pagination mixin is enabled for the group member list
page, but it doesn't appear that the `paginate_by` property was removed
at some point in the past few years. Adding that back and setting the
property to list 40 members, which means most groups should see most if
not all of their member list on one page while still keeping things
performant for groups with thousands of members.