Change the batch size of UserThread creation
When a message is sent a worker creates 1 record in a many-to-many
through table (the UserThread table) for every member in the group
eligible to receive that message. Previously this would be done in one
very large INSERT query. Creating 3000+ rows in 1 table that is actively
being read from by users in 1 query is a bit intensive on the database
and was sometimes failing to create all the records. By changing things
so that these inserts are done in batches of 250 we increase the number
of queries involved in sending a message but potentially reduce the
load on the popular UserThread table.